DTV: revert to Hertz instead of kilohertz unit (fixes #4981)
At least the ISDB-T standard export version uses sub-kHz precision
(never mind that one band occupies several megahertz in width).

+++ modules/access/dtv/dtv.h
@@ -61,10 +61,10 @@ int dvb_set_dvbc (dvb_device_t *, uint32_t freq, const char *mod,
                   uint32_t srate, uint32_t fec);
 
 /* DVB-S */
-int dvb_set_dvbs (dvb_device_t *, uint32_t freq, uint32_t srate, uint32_t fec);
-int dvb_set_dvbs2 (dvb_device_t *, uint32_t freq, const char *mod,
+int dvb_set_dvbs (dvb_device_t *, uint64_t freq, uint32_t srate, uint32_t fec);
+int dvb_set_dvbs2 (dvb_device_t *, uint64_t freq, const char *mod,
                    uint32_t srate, uint32_t fec, int pilot, int rolloff);
-int dvb_set_sec (dvb_device_t *, uint32_t freq, char pol,
+int dvb_set_sec (dvb_device_t *, uint64_t freq, char pol,
                  uint32_t lowf, uint32_t highf, uint32_t switchf);
 
 /* DVB-T */
